Switch to Canvg for client-side PNG rendering (#31)
This is a huge commit that changes a ton of things. Mainly, Lob is now enforcing limits on its API, so we can no longer send huge HTML files and have them render it for us. So, we now use Canvg to do all rendering (and SVG text wrapping...) client-side. There's a bunch of lints, tooling changes, switching to JSX, etc.